requirement.txt
flask
google-genai
gradio
python-dotenv